将附件相关dws迁移到oa.go中，并补充skill描述 to #666
